update: separated styles
