+++ src/components/ParcelCarrier/ParcelCarrier.tsx
@@ -1,7 +1,8 @@
+import "./parcelCarrierStyles.css";
 import React from "react";
-import { Parcel } from "../contracts";
-const wheel = require("../assets/icons/wheel.svg").default;
-const phone = require("../assets/icons/phone.svg").default;
+import { Parcel } from "../../contracts";
+const wheel = require("../../assets/icons/wheel.svg").default;
+const phone = require("../../assets/icons/phone.svg").default;
 
 const ParcelCarrier: React.FC<{ parcel: Parcel }> = ({ parcel }) => {
   return (
