+++ src/components/ParcelCarrier/parcelCarrierStyles.css
@@ -1,0 +1,39 @@
+.carrier-wrap {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  background-color: var(--white);
+  box-shadow: 1px 1px 8px -2px var(--secondary-text-color);
+}
+
+.wheel-icon {
+  padding: 8px;
+  border-radius: 8px;
+  background-color: var(--secondary);
+  color: var(--primary);
+}
+
+.carrier-wrap {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  border-radius: 16px;
+  padding: 16px;
+}
+
+.carrier-name {
+  font-size: 16px;
+  font-weight: 600;
+}
+
+.carrier-tel {
+  font-size: 14px;
+  font-weight: 600;
+  color: var(--secondary-text-color);
+}
+
+.tel-img {
+  padding: 8px;
+  border-radius: 50%;
+  background-color: var(--primary);
+}
