+++ src/components/ParcelTitle/parcelTitleStyles.css
@@ -1,0 +1,47 @@
+.parcelTitleWrap {
+  display: flex;
+  justify-content: flex-start;
+  gap: 20px;
+  margin-bottom: 20px;
+}
+
+.parcel-title {
+  font-size: 24px;
+  font-weight: 900;
+}
+
+.title-descr {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
+
+.info-wrap {
+  display: flex;
+  align-items: center;
+}
+
+.parcel-img {
+  width: 40px;
+  height: 40px;
+  padding: 12px;
+  border-radius: 12px;
+  background-color: var(--secondary);
+  color: var(--primary);
+}
+
+.status {
+  font-size: 14px;
+  font-weight: 500;
+  background-color: var(--blue-label);
+  padding: 4px 12px;
+  border-radius: 12px;
+  margin-right: 12px;
+  color: var(--white);
+}
+
+.price {
+  font-size: 18px;
+  font-weight: 600;
+  color: var(--secondary-text-color);
+}
