+++ src/parcelStyles.css
@@ -65,147 +65,3 @@ a {
   max-width: 335px;
   margin: 0 auto;
 }
-
-.parcelTitleWrap {
-  display: flex;
-  justify-content: flex-start;
-  gap: 20px;
-  margin-bottom: 20px;
-}
-
-.parcel-title {
-  font-size: 24px;
-  font-weight: 900;
-}
-
-.title-descr {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-}
-
-.info-wrap {
-  display: flex;
-  align-items: center;
-}
-
-.parcel-img {
-  width: 40px;
-  height: 40px;
-  padding: 12px;
-  border-radius: 12px;
-  background-color: var(--secondary);
-  color: var(--primary);
-}
-
-.status {
-  font-size: 14px;
-  font-weight: 500;
-  background-color: var(--blue-label);
-  padding: 4px 12px;
-  border-radius: 12px;
-  margin-right: 12px;
-  color: var(--white);
-}
-
-.price {
-  font-size: 18px;
-  font-weight: 600;
-  color: var(--secondary-text-color);
-}
-
-/* -----------CENTER------------- */
-
-.center-wrap {
-  padding: 16px;
-  border-radius: 16px;
-  background-color: var(--white);
-  box-shadow: 1px 1px 8px -2px var(--secondary-text-color);
-  margin-bottom: 20px;
-}
-
-.adress-wrap {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.adress-wrap:first-child {
-  margin-bottom: 14px;
-}
-
-.adress-city {
-  font-size: 16px;
-  font-weight: 600;
-}
-
-.adress-street {
-  font-size: 12px;
-  font-weight: 500;
-  color: var(--secondary-text-color);
-}
-
-.date-wrap {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-top: 20px;
-  position: relative;
-}
-
-.date-wrap::before {
-  content: "";
-  position: absolute;
-  top: -9px;
-  left: -16px;
-  width: 335px;
-  height: 1px;
-  background-color: var(--secondary-text-color);
-}
-
-.date-time {
-  font-size: 14px;
-  font-weight: 600;
-}
-
-/* -------------Carrier------- */
-
-.carrier-wrap {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  background-color: var(--white);
-  box-shadow: 1px 1px 8px -2px var(--secondary-text-color);
-}
-
-.wheel-icon {
-  padding: 8px;
-  border-radius: 8px;
-  background-color: var(--secondary);
-  color: var(--primary);
-}
-
-.carrier-wrap {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  border-radius: 16px;
-  padding: 16px;
-}
-
-.carrier-name {
-  font-size: 16px;
-  font-weight: 600;
-}
-
-.carrier-tel {
-  font-size: 14px;
-  font-weight: 600;
-  color: var(--secondary-text-color);
-}
-
-.tel-img {
-  padding: 8px;
-  border-radius: 50%;
-  background-color: var(--primary);
-}
